+++ packages/sqflite/example/lib/src/expect.dart
@@ -5,7 +5,6 @@
 import 'dart:async';
 
 import 'package:matcher/matcher.dart';
-import 'package:meta/meta.dart';
 
 /// An exception thrown when a test assertion fails.
 class TestFailure {
@@ -117,7 +116,6 @@ FutureOr _expect(actual, matcher,
 
 /// Convenience method for throwing a new [TestFailure] with the provided
 /// [message].
-@alwaysThrows
 void fail([String? message]) => throw TestFailure(message ?? 'should fail');
 
 /// index text helper.