+++ packages/tizen_audio_manager/example/lib/main.dart
@@ -14,7 +14,7 @@ void main() {
 }
 
 class MyApp extends StatelessWidget {
-  const MyApp({Key? key}) : super(key: key);
+  const MyApp({super.key});
 
   @override
   Widget build(BuildContext context) {
@@ -27,7 +27,7 @@ class MyApp extends StatelessWidget {
 }
 
 class VolumeControlScreen extends StatefulWidget {
-  const VolumeControlScreen({Key? key}) : super(key: key);
+  const VolumeControlScreen({super.key});
 
   @override
   State<StatefulWidget> createState() => _VolumeControlScreenState();