+++ packages/video_player_videohole/CHANGELOG.md
@@ -1,3 +1,7 @@
+## NEXT
+
+* Remove use of the deprecated API `WidgetsBinding.instance.window`.
+
 ## 0.1.1
 
 * Update pigeon to 10.0.0.