+++ packages/video_player_videohole/CHANGELOG.md
@@ -1,6 +1,7 @@
-## NEXT
+## 0.1.2
 
 * Increase the minimum Flutter version to 3.3.
+* Fix web build
 
 ## 0.1.1